renamed modules (devtool -> hook)
