+++ src/isomorphic/classic/types/checkReactTypeSpec.js
@@ -17,7 +17,7 @@ var ReactPropTypesSecret = require('ReactPropTypesSecret');
 var invariant = require('invariant');
 var warning = require('warning');
 
-var ReactComponentTreeDevtool;
+var ReactComponentTreeHook;
 
 if (
   typeof process !== 'undefined' &&
@@ -29,7 +29,7 @@ if (
   // https://github.com/facebook/react/issues/7240
   // Remove the inline requires when we don't need them anymore:
   // https://github.com/facebook/react/pull/7178
-  ReactComponentTreeDevtool = require('ReactComponentTreeDevtool')
+  ReactComponentTreeHook = require('ReactComponentTreeHook')
 }
 
 var loggedTypeFailures = {};
@@ -88,13 +88,13 @@ function checkReactTypeSpec(typeSpecs, values, location, componentName, element,
         var componentStackInfo = '';
 
         if (__DEV__) {
-          if (!ReactComponentTreeDevtool) {
-            ReactComponentTreeDevtool = require('ReactComponentTreeDevtool');
+          if (!ReactComponentTreeHook) {
+            ReactComponentTreeHook = require('ReactComponentTreeHook');
           }
           if (debugID !== null) {
-            componentStackInfo = ReactComponentTreeDevtool.getStackAddendumByID(debugID);
+            componentStackInfo = ReactComponentTreeHook.getStackAddendumByID(debugID);
           } else if (element !== null) {
-            componentStackInfo = ReactComponentTreeDevtool.getCurrentStackAddendum(element);
+            componentStackInfo = ReactComponentTreeHook.getCurrentStackAddendum(element);
           }
         }
